added the correct code for the intersection
